+++ src/MLPnPsolver.cpp
@@ -768,7 +768,7 @@ namespace ORB_SLAM3 {
 
         Eigen::MatrixXd jacs(2, 6);
 
-        for (int i = 0; i < pts.size(); ++i)
+        for (size_t i = 0; i < pts.size(); ++i)
         {
             Eigen::Vector3d ptCam = R*pts[i] + T;
             ptCam /= ptCam.norm();
